# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Q-Trust API
  version: "2.2.0"
  description: |
    Q-Trust is a post-quantum trust evaluation and attestation platform built on
    Base (Sepolia / Mainnet). The API provides:

    - **Assets**: register and verify on-chain CBOM (Cryptographic Bill of Materials) hashes.
    - **Orgs**: query organization summaries, assets, migrations, and audits.
    - **Attestations**: vendor product attestation via direct admin writes or
      EIP-712 gasless relay (no wallet needed).
    - **Migrations**: record and track post-quantum migration progress.
    - **Trust Evaluation**: policy-based PQC readiness scoring.
    - **Credentials**: W3C VC v2.0 issuance and fail-closed cryptographic verification (Ed25519Signature2020), byte-compatible with the Python SDK.
    - **Revocation / Policy / Schema / Trust-Anchor**: on-chain contract queries.
    - **Webhooks**: Redis-backed event subscriptions.
    - **Legacy**: pre-v1 alias routes (deprecated, sunset 2026-12-31).

    **Protected operations** require a valid `X-Api-Key` header in production,
    including scans, GPU analysis, write, relay, planner, evidence creation,
    webhook management, and metrics.
    **Gasless relay routes** (`/v1/relay/*`) let vendors sign EIP-712 typed data off-chain;
    the relayer submits the transaction so vendors never need gas tokens.

  contact:
    name: Q-Trust Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001
...
paths:
...
  /v1/webhooks/subscribe:
    post:
      operationId: subscribeWebhook
      tags: [webhooks]
      security:
        - ApiKeyAuth: []
      summary: Subscribe to events
      description: |
        Creates a webhook subscription in Redis. Subscribes to the given event
        types (or `*` for all events).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscribeRequest"
      responses:
        "200":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscribeResponse"
        "400":
          description: Missing address or url
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Redis unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/webhooks/unsubscribe:
    post:
      operationId: unsubscribeWebhook
      tags: [webhooks]
      security:
        - ApiKeyAuth: []
      summary: Unsubscribe from events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookUnsubscribeRequest"
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookUnsubscribeResponse"
        "503":
          description: Redis unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/webhooks/subscribers:
    get:
      operationId: listSubscribers
      tags: [webhooks]
      summary: List webhook subscribers
      description: Returns all active webhook subscriptions grouped by event type.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Subscriber list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscribersResponse"

  # ── Health aliases ─────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Admin API key required for write and credential-issuance endpoints.
        Configured via `QTRUST_API_KEYS` (comma-separated). Enforced only in
        production when keys are configured.
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    WebhookSubscribeRequest:
      type: object
      required: [address, url]
      properties:
        address:
          type: string
          description: On-chain address to watch
        url:
          type: string
          format: uri
          description: Webhook delivery URL
        secret:
          type: string
          description: Optional shared secret for HMAC verification
        events:
          type: array
          items:
            type: string
          description: "Event types to subscribe to (default: [*] for all)"
          default:
            - "*"
    WebhookSubscribeResponse:
      type: object
      properties:
        subscribed:
          type: boolean
        subscriber:
          type: object
          properties:
            address:
              type: string
            url:
              type: string
            events:
              type: array
              items:
                type: string
            secret:
              type: string
              description: Masked secret
    WebhookUnsubscribeRequest:
      type: object
      required: [address, url]
      properties:
        address:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
          description: "Event types to unsubscribe from (default: [*])"
    WebhookUnsubscribeResponse:
      type: object
      properties:
        unsubscribed:
          type: boolean
        removed:
          type: integer
          description: Number of subscriptions removed
    WebhookSubscribersResponse:
      type: object
      properties:
        subscribers:
          type: array
          items:
            type: object
            properties:
              event:
                type: string
                description: Event type (or "*" for wildcard)
              count:
                type: integer
                description: Number of subscribers for this event