# backend/src/routes/webhooks.ts
import type { FastifyInstance } from "fastify";
import { createHash } from "node:crypto";
import { Redis } from "ioredis";
import { requireApiKey } from "../middleware/auth.js";
import { isPublicHttpsUrl } from "../services/webhook.js"; // REG-19
import { encryptSecret, decryptSecret } from "../services/secret-box.js";
import { setSubscriberResolver } from "../services/webhook.js";
import { isValidAddress } from "../config.js";
import { WebhookSubscribeSchema, WebhookUnsubscribeSchema } from "../schemas/index.js";

export async function registerWebhookRoutes(app: FastifyInstance, redis: Redis | null): Promise<void> {
  const redisReady = (): boolean => redis?.status === "ready";

  app.post("/v1/webhooks/subscribe", {
    preHandler: requireApiKey,
    config: { rateLimit: { max: 20, timeWindow: "1 minute" } },
    schema: { body: WebhookSubscribeSchema },
  }, async (request, reply) => {
    const { address, url, secret, events } = request.body as { address: string; url: string; secret?: string; events?: string[] };
    if (!address || !url) {
      return reply.status(400).send({ error: "address and url are required" });
    }
    if (!isValidAddress(address)) {
      return reply.status(400).send({ error: "Invalid address format" });
    }
    try { new URL(url); } catch { return reply.status(400).send({ error: "Invalid url format" }); }
    if (!isPublicHttpsUrl(url)) {
      return reply.status(400).send({ error: "URL must be public HTTPS (no localhost/private)" });
    }
    if (!redis || !redisReady()) {
      return reply.status(503).send({ error: "Redis unavailable — webhook service not ready" });
    }
    const eventList = events && events.length ? events : ["*"];
    let stored: string;
    try {
      stored = JSON.stringify({
        url, address,
        secret: encryptSecret(secret ?? "", () => app.log.warn("QTRUST_WEBHOOK_ENC_KEY not set — webhook secrets are stored UNENCRYPTED in Redis")),
      });
      for (const event of eventList) {
        const key = event === "*" ? "subscribers:*" : `subscribers:${event}`;
        await redis.sadd(key, stored);
      }
    } catch (err) {
      request.log.warn({ err }, "Webhook subscription storage failed");
      return reply.status(503).send({ error: "Webhook subscription storage unavailable" });
    }
    return { subscribed: true, subscriber: { address, url, events: eventList, secret: secret ? "•••" : "" } };
  });

  app.post("/v1/webhooks/unsubscribe", {
    preHandler: requireApiKey,
    config: { rateLimit: { max: 20, timeWindow: "1 minute" } },
    schema: { body: WebhookUnsubscribeSchema },
  }, async (request, reply) => {
    const { address, url, events } = request.body as { address: string; url: string; events?: string[] };
    if (!redis || !redisReady()) {
      return reply.status(503).send({ error: "Redis unavailable" });
    }
    if (!address || !url) {
      return reply.status(400).send({ error: "address and url are required" });
    }
    const eventList = events && events.length ? events : ["*"];
    let removed = 0;
    try {
      for (const event of eventList) {
        const key = event === "*" ? "subscribers:*" : `subscribers:${event}`;
        const records = await redis.smembers(key);
        for (const raw of records) {
          try {
            const parsed = JSON.parse(raw) as { url?: string; address?: string };
            if (parsed.url === url && parsed.address === address) {
              removed += await redis.srem(key, raw);
            }
          } catch { continue; }
        }
      }
    } catch (err) {
      request.log.warn({ err }, "Webhook unsubscription storage failed");
      return reply.status(503).send({ error: "Webhook storage unavailable" });
    }
    return { unsubscribed: true, removed };
  });

  app.get("/v1/webhooks/subscribers", { preHandler: requireApiKey }, async () => {
    if (!redis || !redisReady()) return { subscribers: [] };
    // REG-20: SCAN not KEYS (blocking O(N) on live Redis)
    const keys: string[] = [];
    let cursor = "0";
    try {
      do {
        const [nextCursor, batch] = await redis.scan(cursor, "MATCH", "subscribers:*", "COUNT", 100);
        cursor = nextCursor;
        keys.push(...batch);
      } while (cursor !== "0");
    } catch (err) {
      return { subscribers: [], error: "Webhook storage unavailable" };
    }
    const byId = new Map<string, { id: string; url: string; events: string[] }>();
    for (const key of keys) {
      const event = key.replace("subscribers:", "");
      const records = await redis.smembers(key);
      for (const raw of records) {
        let url = "";
        try {
          const parsed = JSON.parse(raw) as { url?: string; address?: string; secret?: string };
          url = typeof parsed.url === "string" ? parsed.url : "";
        } catch { continue; }
        if (!url) continue;
        const id = createHash("sha256").update(url).digest("hex").slice(0, 16);
        const existing = byId.get(id);
        if (existing) {
          if (!existing.events.includes(event)) existing.events.push(event);
        } else {
          byId.set(id, { id, url, events: [event] });
        }
      }
    }
    return { subscribers: Array.from(byId.values()) };
  });

  if (redis) {
    setSubscriberResolver(async (eventType: string) => {
      const keys = [`subscribers:${eventType}`, "subscribers:*"];
      const seen = new Set<string>();
      const out: Array<{ url: string; secret?: string }> = [];
      for (const key of keys) {
        const records = await redis!.smembers(key);
        for (const raw of records) {
          try {
            const parsed = JSON.parse(raw) as { url?: string; address?: string; secret?: string };
            if (!parsed.url || seen.has(`${parsed.address ?? ""}|${parsed.url}`)) continue;
            seen.add(`${parsed.address ?? ""}|${parsed.url}`);
            let secret: string | undefined;
            try { secret = parsed.secret ? decryptSecret(parsed.secret) : undefined; } catch { console.warn(`Webhook subscriber ${parsed.url}: undecryptable secret — delivering unsigned`); }
            out.push({ url: parsed.url, secret });
          } catch { continue; }
        }
      }
      return out;
    });
  }
}